fix: small changes on general styling, moved styling to shared. updated for x-large screens
