+++ src/tjenesteomrader/DigitalTjeneste/index.tsx
@@ -1,7 +1,7 @@
 import EmojiList from '../components/emojiList/emojiList';
 import HeaderBackground from '../images/headerBackground/headerBackground';
 
-import style from '../index.module.css';
+import style from '../shared/index.module.css';
 import TwoRows from '../components/twoRows/twoRows';
 import BlobText from '../components/blobText/blobText';
 import BottomText from '../components/bottomText/bottomText';
