+++ src/tjenesteomrader/images/footer/footer.module.css
@@ -1,5 +1,4 @@
 .footer {
   width: 100%;
   height: auto;
-  position: absolute;
 }
