+++ src/tjenesteomrader/strategi/index.tsx
@@ -3,7 +3,7 @@ import BottomText from '../components/bottomText/bottomText';
 import EmojiList from '../components/emojiList/emojiList';
 import TwoRows from '../components/twoRows/twoRows';
 import HeaderBackground from '../images/headerBackground/headerBackground';
-import style from '../index.module.css';
+import style from '../shared/index.module.css';
 
 const Strategi = () => {
   return (
